requirements.txt
# For Windows
streamlit
tensorflow
opencv-python
pillow
numpy
pathlib

# For my M4 MBP Tahoe 26.2
# numpy==1.26.4
# tensorflow==2.15.0
# tensorflow-metal
# opencv-python==4.8.1.78


